requirements.txt
git+https://github.com/cedadev/django-haystack-elasticsearch.git
git+https://github.com/cedadev/ceda-elasticsearch-tools.git
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.6
django-bootstrap-datepicker-plus==3.0.5
django-classy-tags==0.9.0
django-cookie-law==2.0.2
django-crispy-forms==1.8.1
django-haystack==2.8.1
git+https://github.com/rsmith013/django-jira-oauth.git
django-model-utils==4.0.0
django-sizefield==1.0.0
django-tables2==2.3.1
django-unused-media==0.2.0
djangorestframework==3.10.3
docopt==0.6.2
docxtpl==0.10.0
elasticsearch==7.6.0
git+https://github.com/cedadev/fwtheme-django.git
git+https://github.com/cedadev/fwtheme-django-ceda-serv.git
git+https://github.com/cedadev/orgtheme-datamad.git
idna==2.9
Jinja2==2.11.2
jira==2.0.0
lxml==4.5.2
MarkupSafe==1.1.1
node==0.9.24
numpy==1.18.1
oauthlib==3.1.0
odict==1.7.0
pandas==1.0.1
pbr==5.4.4
plumber==1.5
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
pytest-runner==5.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.14.0
sqlparse==0.3.0
tqdm==4.43.0
urllib3==1.25.8
zope.component==4.5
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==4.2.0
zope.interface==4.6.0
zope.lifecycleevent==4.3
zope.proxy==4.3.2
